+++ src/compression/snappy.rs
@@ -7,7 +7,7 @@ use crate::protocol::buf::{ByteBuf, ByteBufMut};
 use super::{Compressor, Decompressor};
 
 /// Kafka variant of the snappy compression algorithm. See
-/// https://github.com/xerial/snappy-java?tab=readme-ov-file#compatibility-notes for notes about
+/// <https://github.com/xerial/snappy-java?tab=readme-ov-file#compatibility-notes> for notes about
 /// the difference from standard snappy.
 /// See [Kafka's broker configuration](https://kafka.apache.org/documentation/#brokerconfigs_compression.type)
 /// for more information about compression.
@@ -139,6 +139,7 @@ mod tests {
         let record = Record {
             transactional: false,
             control: false,
+            delete_horizon: false,
             partition_leader_epoch: 0,
             producer_id: 0,
             producer_epoch: 0,
@@ -201,6 +202,7 @@ mod tests {
         let expected_record = Record {
             transactional: false,
             control: false,
+            delete_horizon: false,
             partition_leader_epoch: 0,
             producer_id: 0,
             producer_epoch: 0,
@@ -245,6 +247,7 @@ mod tests {
         let expected_record = Record {
             transactional: false,
             control: false,
+            delete_horizon: false,
             partition_leader_epoch: 0,
             producer_id: 0,
             producer_epoch: 0,